Add ability to control bones per vertex
If any vertex is affected by  more than the maximum number of
bones, the least important vertex weights are removed and the
remaining vertex weights are renormalized so that the weights
still sum up to 1. The default bone weight (bones per vertex)
limit is 4.

Added flag DisableBones, that is equivalent to setBonesPerVertex(0)

A lot of variables were hidden to private scope