Context: fix static initialization order problem
Since Context::addOnDestroyListener may be called in
a static initializer, we can encounter a static
initialization problem. The corresponding
tulz::Subject<Context*> may not have been initialized
yet, which can lead to a so-called static initialization
disaster. To prevent this from happening, the Construct On
First Use Idiom was used (the static object was wrapped
inside a static function).

Read more here:
  https://isocpp.org/wiki/faq/ctors#static-init-order